Add spline slowmotion method

+++ src/spline.h
@@ -10,17 +10,17 @@ class Spline{
 	public:
 		Spline();
 		Spline(int n);
-		Spline(const vector<double> &y);
-		void recalcular(const vector<double> &y);
+		Spline(const vector<int> &y);
+		void recalcular(const vector<int> &y);
 		double evaluar(double x);
 	private:
-		vector<double> a;
+		vector<int> a;
 		vector<double> b;
 		vector<double> c;
 		vector<double> d;
 		vector<vector<double> > factorizacionLU;
 		void generarSistema(int n);
-		void generarSpline(const vector<double> &y);
+		void generarSpline(const vector<int> &y);
 };
 
 #endif // SPLINE_H_INCLUDED
